Move `run_concurrently` from `tests/utils` to `src/oscar/test/utils`
This change will allow to import `run_concurrently` helper for test
purposes in other libraries.
